requirements.txt
alembic==1.15.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.4.22
cffi==2.0.0
click==8.3.3
colorama==0.4.6
cryptography==44.0.3
dnspython==2.8.0
ecdsa==0.19.2
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
Mako==1.3.11
MarkupSafe==3.0.3
passlib==1.7.4
pillow==11.2.1
pyasn1==0.6.3
pycparser==3.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.34.2
watchfiles==1.1.1
websockets==16.0
